requirements.txt
discord
PyNaCl
ruamel.yaml
requests
GitPython